[TASK] Fixes for WallsService to fix issues

+++ Classes/Service/WallsService.php
@@ -20,7 +20,6 @@ use TYPO3\CMS\Core\Core\Environment;
 use TYPO3\CMS\Core\Registry;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Core\Utility\PathUtility;
-use TYPO3\CMS\Core\Utility\StringUtility;
 use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
 
 /**
@@ -29,16 +28,9 @@ use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
 class WallsService
 {
     /**
-     * @var string
-     */
-    protected $targetDirectory = 'typo3temp/assets/walls_io_proxy';
-
-    /**
      * Fields to get from the API
-     *
-     * @var array
      */
-    protected $fields = [
+    protected array $fields = [
         'id',
         'comment',
         'type',
@@ -53,15 +45,9 @@ class WallsService
         'post_link',
     ];
 
-    /**
-     * @var Registry
-     */
-    protected $registry;
-
-    /**
-     * @var WallsIoClient
-     */
-    protected $client;
+    protected string $targetDirectory = 'typo3temp/assets/walls_io_proxy';
+    protected Registry $registry;
+    protected WallsIoClient $client;
 
     public function __construct(Registry $registry, WallsIoClient $client)
     {
@@ -223,7 +209,7 @@ class WallsService
 
         if (
             array_key_exists('external_image', $post)
-            && StringUtility::beginsWith((string)$post['external_image'], 'http')
+            && str_starts_with((string)$post['external_image'], 'http')
         ) {
             $post['external_image'] = $this->cacheExternalResources(
                 $post['external_image'],
@@ -233,7 +219,7 @@ class WallsService
 
         if (
             array_key_exists('post_image', $post)
-            && StringUtility::beginsWith((string)$post['post_image'], 'http')
+            && str_starts_with((string)$post['post_image'], 'http')
         ) {
             $post['post_image'] = $this->cacheExternalResources(
                 $post['post_image'],
@@ -252,7 +238,7 @@ class WallsService
                 && is_array($matches['src'])
             ) {
                 foreach ($matches['src'] as $uri) {
-                    if (StringUtility::beginsWith($uri, 'http')) {
+                    if (str_starts_with($uri, 'http')) {
                         $post['comment'] = str_replace(
                             $matches['src'],
                             $this->cacheExternalResources($uri, $pluginConfiguration->getRecordUid()),
@@ -270,6 +256,7 @@ class WallsService
     protected function cacheExternalResources(string $resource, int $contentRecordUid): string
     {
         $pathParts = GeneralUtility::split_fileref(parse_url($resource, PHP_URL_PATH));
+
         $filePath = sprintf(
             '%s%s.%s',
             $this->getTargetDirectory($contentRecordUid),
@@ -277,7 +264,7 @@ class WallsService
             $pathParts['fileext']
         );
 
-        if (!file_exists($filePath)) {
+        if (!file_exists($filePath) && GeneralUtility::getUrl($resource)) {
             GeneralUtility::writeFile($filePath, GeneralUtility::getUrl($resource));
         }
